Precompile on rank-0 first in run_moment_kinetics.jl
When running from the command line using the `run_moment_kinetics.jl`
script, call `using moment_kinetics` first on rank-0, and only after
that has completed call it on all other ranks. Hopefully this avoids
precompilation clashes (when multiple MPI ranks try to precompile the
same package simultaneously) which might corrupt files. It does not
solve the clash issue for importing MPI, because `using MPI` is needed
to get the MPI commands to synchronize between rank-0 and the other
processes, but hopefully minimises the chance of problems.