Remove runtime plots
It is inconvenient to provide the option for runtime plots when
splitting the post-processing functionality into separate packages. The
option was not used much anyway. It should be possible to add an
'extension' (with Julia-1.9.x) that provides the option if a
post-processing module is loaded, if we want to add it back in future.