Remove brackets from generic function definitions
The 'generic' definitions are intended only to define a name that can be
imported by each implementation to add an implementation-specific
method. If they are defined with brackets after the name (as several
were previously) then there is a zero-argument method of the function
that does nothing, which might sometimes be confusing. Therefore remove
those brackets.