Update CI tests for reorganised package structure
Also bumps the Julia version used for the CI to 1.9, as this is required
in order to use package extensions.

Also adds NCDatasets to test dependencies. This will require/install
NCDatasets when running `Pkg.test()`, but it should still be possible to
run the tests without NCDatasets by just doing
`include("moment_kinetics/test/runtests.jl")`.