Modernise test dependency specification, remove unneeded dependencies
The 'modern' way to specify extra dependencies for the tests in Julia is
using `test` in the `[targets]` section of the top-level Project.toml.
The old method of using an extra `test/Project.toml` is described as
'the old method' which, while it will continue to be supported in
Julia-1.x, seems to be out of favour.

The main point of this update though is to remove extra dependencies
from the tests, which are not needed.