Allow tests to run when optional dependencies are not installed
NCDatasets is now an optional dependency (also Symbolics, but that is
not used in the test suite). To allow the tests to run when NCDatasets
is not installed, the tests that use NetCDF now fall back to HDF5 if
NetCDF is not available.

The fall-back can be disabled explicitly by passing the
--force-optional-dependencies command line flag, to ensure that the CI
tests do actually use NetCDF I/O, and will error if it is not available.