When submitting batch jobs, check system image is newer than source code
If the system image is not newer than the source code, probably the
source code has been modified since the system image was compiled. The
changes to the code will not take effect until the system image is
re-compiled, so print a warning advising the user to re-compile the
system image. We do not error (so leave it up to the user to cancel any
submitted jobs) because sometimes it might be OK to use the 'older'
system image (e.g. if changes were made and then undone, or if the user
really wants to run with the older code).