Remove no-longer-necessary force exit from machine_setup_stage_two.jl
The MPI and HDF5 setup is no longer done in the
machine_setup_stage_two.jl script. It is also not really expected that
the script will be run other than from machine_setup.sh. So the message
saying Julia is exiting and the forced exit by calling `exit(0)` are not
really necessary.