Remove `@debug_error_stop_all` debugging macro
The error handling in `run_moment_kinetics()` now calls `MPI.Abort()` in
any case if there is an error. The implementation of
`@debug_error_stop_all` is a bit complicated already, and does not work
with 'anyv' regions. Seems better to just remove it now rather than try
to make it work with 'anyv'.