Collision operator buffer arrays are shared on 'anyv subblocks'
Creating buffer arrays for the Fokker Planck collision operator as
shared-memory arrays that are only shared between the processes on a
single 'anyv-subblock' (not the whole 'block' as used by the rest of
`moment_kinetics`) avoids having to create 'views' of larger
shared-memory arrays in order to get a unique buffer for each subblock.
Creating the views seemed to have a performance penalty (of ~10% of the
total collision operator evaluation time in one test case).