Allow communicator used for shared-memory arrays to be chosen
For use in an 'anyv' region, it is useful to be able to create
shared-memory arrays that are shared only between the processes in the
same 'anyv subblock'. To allow this, add an optional argument to
`allocate_shared()` that allows the MPI communicator to be specified (by
default `comm_block[]` is used).