Better algorithm to choose processor splits for shared-mem parallelism
Sometimes, for some combinations of dimensions, it might be advantageous
to use factors of a slightly smaller number than the total block size,
e.g. if using 15 processes to split 2 dimensions into [5,3] chunks is
more efficient than using 16 processes to split the dimensions into
[4,4], [2,8], or [8,2] chunks. New algorithm allows this.

Also use the 'maximum work' (i.e. maximum number of grid points
belonging to a single process) as the quantity to minimise, rather than
the 'load balance'. 'Maximum work' is what we actually care about, and
it is simpler to deal with, and it avoids needing hacks to deal with
cases where the 'minimum work' is 0 (because there are too many
processes assigned to some dimension).