Skip enforcing first-moment constraint in electron wall bc
The constraint should be approximately enforced by the way the cutoff is
calculated, so the distribution function cannot drift far from the
constraint being satisfied. The error will be small in spatial
discretization (related to the spacing of the vpa grid), and should also
be constrained by the constraints being applied at neighbouring grid
points. Trying to apply the constraint might interfere with the symmetry
of g(wpa) around vpa=0 (between -v_cutoff and +v_cutoff).