Cut off g_e at sheath with tanh() instead of sharp cut-off
One significant advantage of this update is that the exact position of
the cut-off is allowed to be in between grid points, so the position
does not jump from grid point to grid point (it is interpolated linearly
in between). The jumps from grid point to grid point prevent (or at
least significantly slow down) convergence to a steady state at long-ish
times.

Not clear if there is a large advantage to the tanh-cutoff approach, as
there is still a lot of ringing in the tails of the distribution
function, but if this change is reverted, the replacement should be a
cutoff that rather than just setting points to zero, scales the last
point so that the integral over the tail gives u_i exactly (to avoid the
jumping-from-grid-point-to-grid-point thing).