Settings for non-crashing simulation
Using the current code, and the input in
`examples/kinetic-electrons/wall+sheath-bc_kinetic.toml`, time advance
of the electron distribution function continues for a long time without
crashing. Convergence to steady state seems to be very slow, maybe not
helped by lots of ringing in the tails of the distribution function.