When using ppar z-diffusion for electrons, zero out end points
If z-diffusion is allowed to update the values of electron ppar at the
end points in z (the sheath entrance points), then there are not enough
boundary conditions and so the simulation becomes numerically unstable.
This instability can be avoided (without having to impose unwanted
boundary conditions) by setting the d2ppar_dz2 array to 0.0 at the end
points. The diffusion is just a numerical hack anyway, which should be
removed once a smooth enough initial condition has been reached, so this
extra hack should be OK (?).