For variables not in `input_dict`, just use options from the top-level
For the `timestep_diagnostic_variables` only the time-indexing options
are relevant, and these can only be set at the top level anyway. The
other options (e.g. `show_element_boundaries`) need to be set as they
are used in some places, but do not actually do anything when the plot
axis is 't'.

Also get `input` in Tuple-handling versions of 1d plot_vs_* and
animate_vs_*, to allow setting element boundaries properly.