Modify grid_spacing for finite element grids
The function `grid_spacing()` is used to calculate `cell_width`, but up
to now `cell_width` is only used for finite difference methods, so we
are free to modify the definition for finite element methods.

This commit makes the definition of `cell_width` (for finite element
methods) half of the distance to the two adjacent grid points, except at
the boundaries where `cell_width` is the (full) distance to the adjacent
grid point (the one next to the boundary).