Write out most recent version of constraints coefficients as diagnostic
It can be useful to know how much of a change the 'moment constraint
corrections' are having to apply. To help diagnose this, store the
values of the coefficients used to apply the constraints (for each
species of ion and neutral, at each spatial point) into arrays in the
`moments` structs, and write out the most recent values at each output.