Looser tols for adaptive timestepping tests to deal with rounding errors
Rounding errors mean that the timestep in adaptive timestep simulations
can change slightly, which changes the results and further changes the
timestep. The difference should be small (less than the tolerances
and/or discretisation error), but makes it more or less impossible to do
regression tests with tolerances ~1e-14 that pass on multiple systems.

Also, don't use `high_precision_error_sum=true` to try to make adaptive
timestepping more consistent. Since the rounding errors change on
different machines anyway, this does not fix the rounding-error issue,
so might as well just test the 'usual' mode of operation.