Pass all moments to electron_energy_equation!() explicitly as arguments
When initializing the electron distribution function and parallel
pressure, want to evolve the energy equation using ion and neutral
moments from the `moments` struct, because `scratch` is being used for
the electron time advance, but when advancing the electron parallel
pressure along with the ion/neutral moments and kinetic equation, need
to get some of the ion and neutral moments from `scratch`. This commit
makes electron_energy_equation!() take all the moments as arguments, so
that the calling site decides which struct they come from.