Fix source moments to be compatible with source_type="energy_source"
Previously, the moments of the source term were introduced in vpa/vz
'advection speed' and in source_terms in terms of
`external_source_amplitude`. The moments were taken assuming that the
source is a Maxwellian with temperature `T_source`. This is not true for
"energy_source" (where the source is defined as roughly 'f_M - f' in
such a way that no particles are introduced). However, the moments of
the source term are already calculated in external_density_source and
external_pressure_source, so these can be used instead to support
more generic source terms.