Set electrons t_params.next_output_time[] when not doing initial output
When not doing initial electron output, the
`t_params.next_output_time[]` was not set correctly, so its value would
be either 0.0, or whatever was left over from the electron
initialisation. This might terminate the kinetic electron update
incorrectly, or otherwise cause errors. Instead set it to `Inf` so that
the output-writing blocks in update_electron_pdf_with_time_advance!()
are never triggered.