Fix handling of points between 0 and coord.grid[1] for Radau elements
For a coordinate with a Radau first element, the region between coord=0
and coord=coord.grid[1] is actually within the first element, and there
will be no points to interpolate to to the left of the first element,
because the coordinate range is 0<coord<infinity. Radau elements
therefore need special handling, so that there is no Maxwellian-like
extrapolation to the left of the first element.