Fix restarting of electron initialisation
Need to remove output times that have already been passed, as having
times less than or equal to the initial time would mess up the I/O
logic. Slight hack to workaround the fact that when `t_params.electron`
is set up, we do not know if we are restarting from an existing
`*.initial_electron.h5` file or not.