Do not do kinetic electron solve for lower-order RK solution
When using adaptive timestepping, we use the RK scheme to calculate a
lower-order approximation to the solution, which is used to estimate the
error (which is then used to update the timestep). To apply boundary
conditions and constraints to the lower-order approximations of the ion
and neutral distribution functions, we use the
`apply_all_bcs_constraints_update_moments!()`
function. This function calls the (expensive!) `update_electron_pdf!()`
function that runs the kinetic electron update, but this kinetic
electron update is not needed for the lower-order approximation, so it
is better to skip it for that step.