Fix handling of processes not included in any 'anyv' sub-block
The algorithm for deciding the size of shared-memory ranges may
sometimes exclude a few processes, if this allows the number of grid
points to be factorised in a more favourable way. When this happens to
an 'anyv' sub-block, some extra special handling (introduced in this
commit) is required to make sure that the excluded processes actually
don't do anything.