Delay overwriting of scratch_electron[1]
We want to have the option to save the state in `scratch_electron[1]`,
in order to calculate steady-state residual diagnostics in
post-processing.  To allow this, we need to start the electron
pseudo-time-step with the state in
`scratch_electron[t_params.electron.n_rk_stages+1]` rather than the
state in `scratch_electron[1]`, and copy that state into
`scratch_electron[1]` at the beginning of each pseudo-time-step.