Estimate timescale for dg_e/dt, use to normalise residual for g_e solve
If we just defined the residual for the electron distribution function
solve to be 'dg/dt=0', then we would be asking the solver (roughly) to
find g such that 'dg/dt<tol' for some tolerance. However dg/dt has some
typical timescale, τ, so 'dg/dt∼g/τ'. It would be inconvenient to have
to define the tolerances taking τ (normalised to the reference sound
crossing time) into account, so instead estimate the relevant timescale
as 'sqrt(me/mi)*z.L', i.e. as the electron thermal crossing time at
reference parameters. We pass this as 'dt' to
electron_kinetic_equation_euler_update!() so that it multiplies the
residual.