Use fixed, small tolerance for Newton iteration in electron sheath bc
Previously was setting the Newton iteration tolerance proportional to
the electron time-solver tolerance, but it is better to use a fixed,
small value (now 1.0e-11). The electron time-solver tolerance might be
set to a large value, in order to set the electron timestep only by CFL
limits, not by truncation error, and anyway the Newton iteration is used
to impose the moment constraints, which should be satisfied 'exactly'.
Even small errors, as they only affect a single grid point (at the
sheath entrance boundary) can cause significant errors in gradients that
are then used elsewhere.