Limit 'sim time' for each electron solve, not number of iterations
When using adaptive timestepping, if the timestep becomes very small it
is very unlikely that the electron solve will converge in a fixed
(small-ish) number of iterations. This makes the electron solve
fragile. It should be more robust to fix the 'sim time' (i.e. normalised
physical time that the simulation is run for) for each electron solve,
which is what is implemented in this commit.