Add 'dense' second derivative matrix to gauss_legendre
For calculating preconditioners, it can be useful to have an explicitly
calculated matrix
```
dense_second_deriv_matrix = inv(mass_matrix) * K_matrix
```
which includes the inverted mass-matrix already. Because of the matrix
inverse, dense_second_deriv_matrix is a dense matrix, so (for
efficiency) should not be used unless absolutely necessary.