Fix 'constant' bc, make it compatible with moment-kinetic mode (?)
'constant' boundary condition may still not be consistent with moment
constraints, but does now use the 'unnormalised' speed to set the
incoming Maxwellian distribution, and the incoming distribution is
normalised by n/vth.

Also fix normalisation - the 1/sqrt(pi) is now included in the
normalisation of f (and in the integration routines) so does not need to
be divided out here.