Recalculate electron preconditioner when Newton solve fails
When the Newton solve fails due to taking too many iterations, it might
be that the preconditioner was too old so recalculating it is useful.
More importantly, when the Newton iteration fails, we decrease dt by a
factor 0.5 - the large change in dt will make the preconditioner bad, so
it needs to be updated.