Remove distinction between periodic_bc_lhs and periodic_bc_rhs
As the input can be assumed to be periodic when using periodic bc, it is
not necessary to zero out the last row of the 'rhs' matrix - it is ok to
use a similar form (1 0 0 ... 0 0 -1) as the last row of the 'lhs'
matrix, because the two non-zero entries will cancel exactly due to the
periodicity. This is convenient because it means we do not need to
distinguish between 'lhs' and 'rhs' matrices - we can use the same
matrix for both functions.