Hard-code the name for the 'stop file'
The name for the 'stop file' used to be changeable in the timestepping
options, but if it was set would have to be set with either an absolute
path or a path relative to where `julia` is running, not relative to the
output directory, so this option wasn't very helpful anyway. Now remove
the option, and just hard code the name to "stop".