Save defaults from `@kwdef structs` in `options`
When an `@kwdef struct` is passed to `set_defaults_and_check_section!()`
so that the defaults are applied by the struct, also load the final
values back into the `options` Dict so that they can be saved to the
output file.

Note that there is no null value in TOML, or in HDF5, so default values
of `nothing` are not allowed, as these could not be saved. Throw a
helpful error if `nothing` is found in options, also add a note in the
docs about this.

Some defaults in [numerical_dissipation], [timestepping],
[electron_timestepping] and [output] sections had to be changed from
`nothing`.