Don't use z-upwinding at distributed-MPI boundaries in Jacobian
Would have expected including upwinding to be a good idea, to be
consistent with `electron_z_advection!()`, but upwinding makes
convergence for the implicit solve in `electron_backward_euler!()` much
slower (~10x more iterations).

Maybe this could indicate that it is more important to have a fully
self-consistent Jacobian inversion for the `electron_vpa_advection()`
part rather than taking half(ish) of the values from one block and the
other half(ish) from the other.