ADI preconditioner
Add a preconditioner for kinetic electrons using a variation on the
'alternating direction implicit' (ADI) method. Terms that couple
velocity space are first solved implicitly (over vpa only for the 1V
case), with terms with z-coupling treated 'explicitly' by being
subtracted from the right-hand-side; then terms that couple in z are
solved implicitly, with terms with v-coupling treated 'explicitly'. The
two steps may be iterated more times if this is needed.