Use OrderedDict or SortedDict everywhere
It is important to have AbstractDicts with a deterministic order of
entries for options, etc., so that when writing them to the output files
with parallel I/O, the writes all happen in exactly the same order,
which is necessary for consistency. Failing to do this can cause hangs
when writing output because the HDF5 cache(s) are inconsistent on
different ranks, which can lead to some ranks calling an MPI function
(e.g. MPI.Bcast!()) when others do not.