Add wall bc to kinetic electron Jacobian
The boundary condition matrix should be added to the Jacobian so that it
acts as a constraint setting the boundary points equal to the values
they would have due to the boundary condition. This can be arranged by
setting the RHS equal to zero, and the rows of the Jacobian
corresponding to the boundary points to (I - M_wall), where I is the
identity and M_wall is the matrix that would generate the perturbations
of the boundary points given the perturbations of the state vector.