Improve interpolation for kinetic electron bc in element containing zero
Previous scheme just interpolated using existing values in whole element
containing zero. This meant that the result of the boundary condition
depended on some points that are overwritten by the boundary condition.
Improve on this by doing special interpolation in the element containing
zero. Instead of using the usual `interpolate_to_grid_1d!()` function,
use `interpolate_symmetric!()` which does an interpolation that is
forced to be symmetric around v_parallel=0, as the interpolation
polynomial is a polynomial in `v_parallel^2` (the interpolating
polynomial is constructed using a Lagrange polynomial method).  The
inputs to the interpolation are now just the function values on grid
points (within the element containing zero) that are not set by the
boundary condition.

Also optimises the interpolation of the points in the elements not
containing zero by restricting the interpolation to just the points
needed for output, instead of interpolating to the full reversed grid.