Make split of integrals at +/-vcut more consistent
Where integrals over the into-the-sheath part of the distribution
function are split into two parts at +/- vcut, modify the way that the
split is done so that the 'part 2' integral between 0 and +/-vcut is
calculated the same way as an integral over the out-from-the-sheath part
of the distribution which is cut off at -/+vcut.

Not sure if this is necessary, but seems nicer to be more consistent.