Narrower cutoff near zero of correction terms for electron bc integral
If the prefactor that sets the correction terms to be proportional to
vpa^2 near vpa=0 is too broad, then it is hard for the correction terms
to fix errors in low moments (e.g. density moment), so making it a bit
narrower reduces the size of the coefficients of the correction terms.