Make kinetic electron bc more robust if a guess for phi_wall is 0
If phi_wall=0, giving vcut=0, then it is not possible to apply moment
constraints, which causes an error due to a singular matrix. This case
is unphysical, so should not ever be a converged solution (vcut=0
corresponds to a fully electron-absorbing sheath). To avoid the problem,
check if vcut==0, and if so set vcut to some small value (we choose the
value at the next grid point after the one closest to 0).

[skip ci]