Fix adaptive-timestep IMEX for kinetic electrons
Can (?) skip the problematic solve, that was run to update the electron
shape function without updating the electron pressure, on the 'explicit
first stage' of ESDIRK schemes. This might have the effect of reducing
the order of accuracy of the scheme somehow, as the qpar_e used for the
'explicit' calculation of the time derivative of electron_ppar is taken
from the most recent implicit solve rather than a solve updated with the
new ion/electron profiles. However, the change is probably small (?) and
at least the solver does run now - it is useful to have an adaptive ion
timestep as it may let the code recover by reducing the ion timestep
when an electron implicit solve fails to converge.