When using adaptive timestep, don't force fixed output times by default
When running kinetic electron simulations, it can cause problems to take
a very short ion timestep. When writing outputs at exactly fixed output
times, this can happen if the previous timestep happened to end just
before the output time. To avoid the very short step default to writing
output at whatever time the end of the timestep is that exceeds the set
output time. There is an option to force the previous behaviour of a
decreased timestep so that output is written exactly at the nominal
output time.