Option to time-evolve the electron shape function
Uses an implicit timestep so that the electron shape function can be
advanced with the ion timestep. Does not test the accuracy of the
timestep for the electron shape function, so that the error control on
the timestepper does not try to resolve fast electron dynamics.

Can be used to find steady state, because then dg_e/dt=0 in the steady
state anyway, so there is no error. When the time-dependence is needed,
dg_e/dt should be a mass-ratio-small term, so maybe is OK anyway?