Fix writing of 'dt' when 'stopnow' is used
When the 'stopnow' file is created to immediately stop a run,
`t_params.dt_before_output[]` was previously not set, because the
`adaptive_timestep_update_t_params!()` function does not know that there
is going to be an output triggered by 'stopnow', so 'dt' was written
incorrectly. Fix this by setting
`t_params.dt_before_output[] = t_params.dt[]` when 'stopnow' is found.