Use explicit pseudotimestep for electrons when ion timestep not IMEX
If the ion timestep is not IMEX, the electron pressure cannot be
advanced implicitly, which is required for a stable implicit electron
timestep. Therefore when not using an IMEX timestep method for ions, use
explicit timestepping for electron pseudotimestepping loop.

Also include a few small fixes for explicit electron timestepping, to
update things that were mildly broken by recent updates.