Micro-optimise a few electron-advance loops
Removes some allocations, and speeds up a few loops. The speed-up for a
given loop is substantial, but these loops were not a large fraction of
the run-time, so the overall effect will be minimal.

Also adds a few extra timers to ensure all functions in
`electron_kinetic_equation_euler_update!()` are timed.