Always catch exception when variable not present in file
This probably either means an old output file, or some variable that is
not always written out (e.g. the PI controller integral terms). Neither
of these is a bug/issue that needs fixing, so makes sense to catch the
error and continue even if `handle_errors = false` is set in the
post-processing input file.