Use moment-time-derivative values instead of substituting by hand
Simplifies code, making fewer places that need to be updated if new
terms are added (now only RHS of kinetic equation and moment equations,
not zdot, wpadot, Fdot coefficients).

Updates expected values in some tests: discretization errors are changed
by the rearrangement of the moment kinetic advection coefficients, and
the tests are generally poorly resolved (in order to be fast), so are
sensitive to this.