Fix electron T equation update when n changes at the same time
We don't currently use a time advance scheme where the electron density
(equal to the ion density) is updated during the same step as the
electron pressure (because explicit and implicit steps are separate even
if the electrons are updated directly on the ion timestep), but if we
did, and also used the temperature equation option for the electrons
(instead of the default pressure equation), then the implementation
would be wrong because it did not distinguish the density at the
beginning and the end of the timestep (compare the implementation of the
parallel velocity update using a momentum equation in force_balance.jl).
This commit fixes this potential error, by passing in separate
electron_density_out and electron_density in to
electron_energy_equation!().