Remove factors of sqrt(pi) from new dimensionless equations
Simpler to write moments, source terms, collisions, etc. without these
factors. They worked better when there was only the simplified CX
operator on the RHS of the kinetic equation. Might as well remove them
while we are redefining the dimensionless variables.

Not sure the sqrt(pi) factors in the new docs were correct before this.