Install NCDatasets in 'debug checks' CI test
Although other tests segfault with NCDatasets, HDF5 fails to precompile
when using OpenMPI_jll unless NCDatasets is also installed. NCDatasets
was originally removed from the 'debug checks' CI test only because it
didn't work for some reason, so makes sense to add it back now that it
works again on the CI server, and for some strange reason helps HDF5 to
precompile.