Update resolution parameters, expected results in Harrison-Thompson test
Increase z-resolution, decrease Lvpa (except in the split3 case),
decrease timestep.

Reduce order of vpa elements to make test more robust.

Switch the test to use `elementwise_isapprox()` instead of `isapprox()`
because `isapprox()` uses `LinearAlgebra.norm()`, which does not divide
out the number of grid points, so that the tolerances would have to be
increased when increasing the grid size.