Use incomplete second deriv matrix to make dissipation precon sparser
The numerical dissipation term in the electron preconditioner is dense
in vpa, because the diffusion matrix is formed as D=M^-1.K. Even though
the mass matrix M and stiffness matrix K are both sparse, the matrix
inverse makes D dense. This denseness is inconvenient for the
preconditioner, but the off-diagonal terms should be small and the
coefficient of the numerical dissipation term should also be small, so
maybe it is OK to just zero them out to make the matrix as sparse as the
first derivative matrix.