Pre-set non-zero pattern of kinetic electron Jacobian matrix
By filling an initial non-zero pattern that covers all possible non-zero
elements of the kinetic electron Jacobian, and then using the new
function `refill_sparse_matrix!()` to update that sparse matrix with the
updated Jacobians, we can ensure that the pattern of non-zero entries
never changes, so we can always use `lu!()` to update the LU
factorisation. This should make the update of the factorisation more
efficient, at the cost of possibly making the `ldiv!()` operation
slightly more expensive because there are a few 'structural non-zeros'
that are actually zero (this happens at the wall boundary points).