Remove upar term in electron_z_advection which is small in sqrt(me/mi)
This term was added when trying to resolve issues with the kinetic
electron sheath boundary condition, but is small in sqrt(me/mi).
Including it is not consistent with the orderings used - to keep this
order would require other terms - and it does actually have a negligible
effect, so makes most sense to remove it.