Hacks: match old full-f initialisation grid before convert to shape func
The redefinition of vth and consequent redefinition of wpa coordinate
mean that to match a simulation from before PR322, the vpa:L input for a
fully-moment-kinetic (evolve_p=true) simulation must be set to sqrt(3)
times the old setting. However the vpa.grid array is used as a
v_parallel (rather than w_parallel) grid to define a 'full-f'
distribution function during initialization (of simulations with wall
boundaries in z), that is then converted to a shape function by
interpolation. The v_parallel grid should only be scaled by sqrt(2), not
sqrt(3) from the old code, making it impossible to exactly match a
simulation from the old code using the new code, without the hacks
introduced in this commit that effectively scale the grid used for the
full-f initialisation by a factor of sqrt(2/3). This hack can probably
be removed once the conversion is verified.