Better initial conditions for recycling-fraction tests
Modifying the `upar_amplitude` setting for neutrals from -1 to +1
results in a more sensible initial profile, with more neutrals going
towards the walls, which hopefully reduces the sensitivity of the test
to numerical noise or rounding errors.

Also removes the numerical dissipation from the split-3 case, as this
also seemed to exaggerate the effect of rounding errors.