Use RK scheme to update PI controller integrals
Previously the integrals were updated as though each timestep was a
forward-Euler timestep, and were also updated to the end of the current
timestep before being used (which seems incorrect - e.g. at the start of
the first timestep t=0, so the integral should be 0 when it is first
used). This commit makes the integrals be handled consistently by the RK
scheme.

[skip ci]