Remove obsolete `debug_dump()` function
The functionality of `debug_dump()` has been mostly replaced by the
`debug_io=true` option, although `debug_dump()` was slighly more
flexible in where it could be called, as it allowed only a selected
subset of variables to be written, where that was useful. However, as
`debug_dump()` required changes to the source code to be used, it was
difficult to test and maintain (so it was probably broken most of the
time) and was less convenient to use.