Update expected results for braginskii_electrons_imex_tests.jl
It is not possible to implement Braginskii closures consistently in a
'1V' system that assumes T_perp=0. The original version (before PR#322)
used T_parallel/p_parallel in place of T/p. The new version introduced
in PR#322 uses the full T=T_parallel/3 and p=p_parallel/3, which allows
the same expressions to be used for 2V mode where T_perp~T_parallel. The
update changes the expected results of
braginskii_electrons_imex_tests.jl. The test would pass with the old
expected results (with the old results converted to the new
dimensionless conventions as they were before this commit) if the dTe_dz
contribution to friction and qpar_e (in the new version) are both
multiplied by 3 to compensate for the difference between T and
T_parallel.