In makie_post_processing, handle animation of variable that is constant
Previously when animating a variable and the minimum and maximum of a
variable were equal, tried to pass them to `ylims!()` anyway, which
caused an error. This commit adds a check that the minimum and maximum
are not equal before calling `ylims!()`, leaving the y-limits at the
default values if they are equal.