Convert 'anyv' parallelisation to 'anysv'
Anticipate that when calculating multi-species collisions, the collision
operators for every species should be calculated together, to be able to
re-use parts of the computation. Therefore rather than 'anyv' regions
inside an outer loop over species, r, and z, we want to have an outer
loop over r and z, containing inner loops over both species and velocity
space - call this an 'anysv' region (for any 's'pecies or 'v'elocity).