Remove unnecessary saving and reloading of r-boundary pdf values
These were only saved and reloaded in case they were used as a radial
boundary condition. However, if they are used as a boundary condition,
then the radial boundary points of the distribution functions will be
set to the relevant values. The boundary values do not need to be saved
separately unless the aim was to do an initial run with some other
radial boundary condition, and then later restart setting the boundary
values to the initial values from the first run - this seems like a very
unlikely use-case, it is more likely that it would be useful to run one
simulation (e.g. a 1D one) to steady state, then restart and fix the
boundary values to the values at the beginning of the restart (which
would be supported by the code as of this commit, but not the previous
version).