Wrap results of build_function() in Symbolics.value()
Not sure why this is necessary, but
`build_function(...; expression=Value{false})` now returns a results
which is a `Num` type from Symbolics.jl. `Num` is not convertible to
Float64, which was causing an error. One way to get around this is to
wrap the function so that the result is passed to `Symbolics.value()`
which extracts the actual value (which in our case is a Float64) from
the `Num`. Not sure this is expected behaviour, so maybe Symbolics.jl
will fix the underlying problem in future, but for now this seems to
work.