consistent naming for update_vpa and update_vz
Change vpa_advection structure to allow for 2D or geometry and completely separate magnetic field strength and mu conservation from the current 1D2V MKE, so that wperp and wpa advection can be kept clean for 1D2V moment kinetics. The flag to allow for geometry has been simply made to be the geometry.input.option == default flag, but we will most likely want a more general flag somewhere in the code eventually.