Cook up inputs to match old 'helical' geometry in MMS report inputs
Implementation of 'helical' geometry has changed, so that the 'old'
inputs resulted in ExB drift being zero (due to the old settings
defining magnetic field to be purely vertical, so v_ExB,z is zero as it
is parallel to B and v_ExB,r is zero as Ez is vertical, so v_ExB,r would
only come from the toroidal E, which is zero - in the old version the
geometry was implemented with a large aspect ratio assumption that is
invalid for these input parameters). Need to cook up a new input that is
effectively the same as the old run - Bz=B/2 and Lz=0.5 (instead of 1),
and tweak 'rhostar' to cancel out the effect of the changed parameters.